@@ -1,14 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <Version>0.2.1</Version>
-        <Description>A modular restartable host built on top of Microsoft.Extensions.Hosting</Description>
-    </PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <Version>0.2.1</Version>
+    <Description>A modular restartable host built on top of Microsoft.Extensions.Hosting</Description>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
 
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Experimental.Collections" Version="1.0.6-e190117-3" />
-      <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Experimental.Collections" Version="1.0.6-e190117-3"/>
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0"/>
+  </ItemGroup>
 
 </Project>
