@@ -1,15 +1,13 @@
-﻿using System.Collections.Generic;
-using System.Reflection;
+﻿using System.Reflection;
 
-namespace Kantaiko.Hosting.Hooks
+namespace Kantaiko.Hosting.Hooks;
+
+public class ApplicationInitializedHook : IHook
 {
-    public class ApplicationInitializedHook : IHook
+    public ApplicationInitializedHook(IReadOnlyList<Assembly> assemblies)
     {
-        public ApplicationInitializedHook(IReadOnlyList<Assembly> assemblies)
-        {
-            Assemblies = assemblies;
-        }
-
-        public IReadOnlyList<Assembly> Assemblies { get; }
+        Assemblies = assemblies;
     }
+
+    public IReadOnlyList<Assembly> Assemblies { get; }
 }
