@@ -1,14 +1,13 @@
 ﻿using Kantaiko.Hosting.Internal;
 
-namespace Kantaiko.Hosting.Host
-{
-    public class HostBuilderConstructionContextProvider : IHostConstructionContextProvider
-    {
-        public ModuleCollection ModuleCollection { get; } = new();
+namespace Kantaiko.Hosting.Host;
 
-        public HostConstructionContext GetHostConstructionContext()
-        {
-            return new(ModuleCollection.ModuleTypes);
-        }
+public class HostBuilderConstructionContextProvider : IHostConstructionContextProvider
+{
+    public ModuleCollection ModuleCollection { get; } = new();
+
+    public HostConstructionContext GetHostConstructionContext()
+    {
+        return new(ModuleCollection.ModuleTypes);
     }
 }
