@@ -1,23 +1,22 @@
 ﻿using Kantaiko.Hosting.Modules;
 
-namespace Kantaiko.Hosting.Host
+namespace Kantaiko.Hosting.Host;
+
+public class ManagedHostBuilder
 {
-    public class ManagedHostBuilder
+    protected string[]? Args { get; }
+
+    public ManagedHostBuilder(string[]? args = null)
     {
-        protected string[]? Args { get; }
+        Args = args;
+    }
 
-        public ManagedHostBuilder(string[]? args = null)
-        {
-            Args = args;
-        }
+    protected HostBuilderConstructionContextProvider ConstructionContextProvider { get; } = new();
 
-        protected HostBuilderConstructionContextProvider ConstructionContextProvider { get; } = new();
+    public IModuleCollection Modules => ConstructionContextProvider.ModuleCollection;
 
-        public IModuleCollection Modules => ConstructionContextProvider.ModuleCollection;
-
-        public virtual IManagedHost Build()
-        {
-            return new ManagedHost(Args, ConstructionContextProvider);
-        }
+    public virtual IManagedHost Build()
+    {
+        return new ManagedHost(Args, ConstructionContextProvider);
     }
 }
