@@ -1,10 +1,7 @@
-﻿using System;
+﻿namespace Kantaiko.Hosting.Exceptions;
 
-namespace Kantaiko.Hosting.Exceptions
+public class ModuleConstructionException : KantaikoHostingException
 {
-    public class ModuleConstructionException : KantaikoHostingException
-    {
-        public ModuleConstructionException(Type moduleType, string message) : base(
-            $"Unable to construct module of type {moduleType.Name}: {message}") { }
-    }
+    public ModuleConstructionException(Type moduleType, string message) : base(
+        $"Unable to construct module of type {moduleType.Name}: {message}") { }
 }
