@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace Kantaiko.Hosting.Host;
 
-namespace Kantaiko.Hosting.Host
+public class HostConstructionContext
 {
-    public class HostConstructionContext
+    public HostConstructionContext(IReadOnlyList<Type>? moduleTypes = null)
     {
-        public HostConstructionContext(IReadOnlyList<Type>? moduleTypes = null)
-        {
-            ModuleTypes = moduleTypes ?? ArraySegment<Type>.Empty;
-        }
-
-        public IReadOnlyList<Type> ModuleTypes { get; }
+        ModuleTypes = moduleTypes ?? ArraySegment<Type>.Empty;
     }
+
+    public IReadOnlyList<Type> ModuleTypes { get; }
 }
