@@ -1,12 +1,9 @@
-﻿using System;
+﻿namespace Kantaiko.Hosting.Modules;
 
-namespace Kantaiko.Hosting.Modules
+[Flags]
+public enum ModuleFlags
 {
-    [Flags]
-    public enum ModuleFlags
-    {
-        None = 0,
-        Library = 1 << 0,
-        Universal = 1 << 1
-    }
+    None = 0,
+    Library = 1 << 0,
+    Universal = 1 << 1
 }
