@@ -1,17 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace Kantaiko.Hosting.Runtime;
 
-namespace Kantaiko.Hosting.Runtime
+internal class RuntimeHostState : IRuntimeHostState
 {
-    internal class RuntimeHostState : IRuntimeHostState
-    {
-        public bool RestartRequested { get; set; }
-        public DateTime RestartRequestedAt { get; set; }
-        public Dictionary<string, object> Properties { get; set; } = new();
+    public bool RestartRequested { get; set; }
+    public DateTime RestartRequestedAt { get; set; }
+    public Dictionary<string, object> Properties { get; set; } = new();
 
-        public bool HostRestarted { get; set; }
-        public bool RestartFailed { get; set; }
-        public DateTime RestartCompletedAt { get; set; }
-        public Exception? StartupException { get; set; }
-    }
+    public bool HostRestarted { get; set; }
+    public bool RestartFailed { get; set; }
+    public DateTime RestartCompletedAt { get; set; }
+    public Exception? StartupException { get; set; }
 }
